+++ tests/CollectorTest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Hkyss\Sonar\Tests;
+namespace hkyss\Sonar\Tests;
 
-use Hkyss\Sonar\Collector;
+use hkyss\Sonar\Collector;
 use PHPUnit\Framework\TestCase;
 
 final class CollectorTest extends TestCase