+++ tests/ConfigTest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Hkyss\Sonar\Tests;
+namespace hkyss\Sonar\Tests;
 
-use Hkyss\Sonar\Config;
+use hkyss\Sonar\Config;
 use PHPUnit\Framework\TestCase;
 
 final class ConfigTest extends TestCase